moved components to the components folder
