+++ src/components/PlayerListCreator.vue
